+++ public/pages/login/index.html
@@ -30,7 +30,7 @@
 					</div>
 
 					<!-- Formulário -->
-					<form id="loginForm" class="w-full">
+					<form action="" id="loginForm" class="w-full">
 						<!-- Campo E-mail -->
 						<div class="flex w-full flex-col items-stretch mt-6 font-normal">
 							<div class="flex items-center min-h-8 gap-1 text-sm leading-loose h-8">
