+++ public/pages/register/index.html
@@ -28,7 +28,7 @@
 
 				<h2 id="titulo-cadastro" class="text-3xl font-semibold text-gray-800 mb-6 text-center">Cadastro</h2>
 
-				<form class="grid grid-cols-1 md:grid-cols-2 gap-6 md:gap-4 md:gap-x-20" id="form-cadastro">
+				<form action="/login" class="grid grid-cols-1 md:grid-cols-2 gap-6 md:gap-4 md:gap-x-20" id="form-cadastro">
 					<div class="text-left">
 						<label for="cpf" class="text-sm text-gray-700">CPF</label><span class="text-error"> *</span>
 						<input id="cpf" type="text" name="cpf" class="w-full p-3 border rounded-md focus:ring-2 focus:ring-secondary outline-none" />
