+++ src/scripts/routes.js
@@ -6,22 +6,20 @@ const ProprietarioAPI = require("./api/Proprietario");
 const loginAPI = require("./api/login");
 const getHorsesAPI = require("./api/apiExample");
 const GerenteAPI = require("./api/Gerente");
-const createHarasAPI = require("./api/createHaras");
+const HarasAPI = require("./api/Haras");
 const getCepAPI = require("./api/getCep");
 const getUsuarioLogadoAPI = require("./api/getUsuarioLogado");
 const deleteFuncionarioAPI = require("./api/deleteFuncionario");
-const getAllHarasAPI = require("./api/GetAllHaras");
 const loginExpiradoAPI = require("./api/loginExpirado");
 
 router.use("/api", ProprietarioAPI);
 router.use("/api/login", loginAPI);
 router.use("/api/exemplo", getHorsesAPI);
 router.use("/api", GerenteAPI);
-router.use("/api/criarHaras", createHarasAPI);
+router.use("/api", HarasAPI);
 router.use("/api/getCep", getCepAPI);
 router.use("/api/getUsuarioLogado", getUsuarioLogadoAPI);
 router.use("/api/deletarFuncionario", deleteFuncionarioAPI);
-router.use("/api/getAllHaras", getAllHarasAPI);
 router.use("/api/loginExpirado", loginExpiradoAPI);
 
 module.exports = router;
